Add support for NFS persisent provisioning
Add roles to setup and teardown the nfs-provisioner. The nfs-provisioner
is configured to retain data and not archive so that cluster restarts
do not dump data. The nfs-provisioner chart is pulled from git so that
we get fixes to enable the pathPattern.

Update the setup playbook to install the nfs-common package on all nodes
which is required for the provisioner.

Update the setup and teardown playbooks to make the nfs-provisioner
available to other deployments (traefik, pihole, etc).

With this change PVCs that specify `spec.storageClassName: nfs-client`
will automatically provision a PV connected to the nfs share. Created
PVs will be named <PVC namespace>/<PVC name>.

Change-Id: I6318c164934465f40db5ad37bf14ee77816c2711