Add role for raspberry pi specific tasks
On raspberry pi there is no need to give up memory to the GPU when it
can instead be used by kublet. The new raspberry pi role is intended
to run tasks that are common to all raspberry pis in the cluster, not
just those on the Turing Pi backplane.

Change-Id: I53e338b2ac7796b19234bc40ff15e630afba9d8a