Fix Traefik dashboard password definition
Traefik was using a static definition for the basic auth middleware that
was defined as part of the example I copied. This fixes that by creating
a temporary file to produce htpasswd encoded output which is then piped
into the secret used for the middleware.

General cleanup of the networking role.

Change-Id: I5c9dc631a83ba4430ae7b80fd4630f946033138d