Setup rtl-sdr driver
Create a new inventory group 'rtl_sdr' for devices that have the sdr
device connected.

Blacklist the DVB drivers that the rtl-sdr is otherwise recognized as so
that we can use it how we want.

Install the rtl-sdr driver package.

Reboot the host if we've made any of these changes.

Change-Id: Ib818c8d78cf5665d5a9dce00ea64116ac26423e5